evidence-engine/pyproject.toml
[project]
name = "ingestion-service"
version = "0.1.0"
description = "File ingestion pipeline for investigations console knowledge graph"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pydantic-settings>=2.7,<3",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "neo4j>=5.27,<6",
    "chromadb>=0.6,<1",
    "redis>=5.2,<6",
    "arq>=0.26,<1",
    "openai>=1.60,<2",
    "PyMuPDF>=1.25,<2",
    "python-docx>=1.1,<2",
    "openpyxl>=3.1,<4",
    "httpx>=0.28,<1",
    "python-multipart>=0.0.18",
    "aiofiles>=24.1,<25",
    "websockets>=14,<15",
    "pyyaml>=6.0,<7",
    "Pillow>=10.0,<11",
    "pytesseract>=0.3.10,<1",
    "lxml>=5.0,<6",
    "markdown-it-py>=3.0,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "httpx>=0.28,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
